Move `utils::content_safe` and dependants into its own module
